[refactor] 폴더 구조 변경으로 인한 import 경로 수정
