+++ src/Pages/TeamPage/Views/FindingTeamDetail.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useParams } from 'react-router-dom';
-import DetailPageBoard from '../../../Components/TeamPage/Contents/DetailPage/DetailPageBoard';
+import DetailPageBoard from '../../../Components/TeamPage/DetailPage/DetailPageBoard';
 
 // 현재 미사용 페이지임.
 function DetailPage() {
   const { id } = useParams();
   const detailList = [
     { title: '활동지역', value: 'area' },
-    { title: '포지션', value: 'position' },
-    { title: '실력수준', value: 'skill' },
+    { title: '모집인원(GK)', value: 'gk_need' },
+    { title: '모집인원(Player)', value: 'player_need' },
     { title: '성별', value: 'gender' },
   ];
   type DataType = {
@@ -18,23 +18,29 @@ function DetailPage() {
     author: string;
     area: string;
     status: string;
-    position: string;
-    skill: string;
     gender: string;
     body: string;
+    gk_need: number;
+    gk: number;
+    player_need: number;
+    player: number;
+    allowRandom: string;
     [key: string]: string | number | undefined;
   };
 
   const dummydata: DataType = {
-    num: 3,
-    title: '나 이민우. 우승팀 들어간다. 불러라.',
-    author: '이민우',
-    area: '부산',
+    num: 1,
+    title: '팀구합니다',
+    author: 'ㄱㅁㅇ',
+    area: '서울',
     status: '미완료',
-    position: '필드플레이어',
-    skill: '프로',
     gender: '남',
-    body: "아이엠 Minu Lee? Isn't this enough?",
+    gk_need: 1,
+    gk: 1,
+    player_need: 4,
+    player: 3,
+    allowRandom: '허용',
+    body: '<p>ㅇㅇㅂㅈㅇㅍ</p><p>ㅂㅈㅇㅍㅂㅈㅇㅍ</p><p><br></p><p>되냐</p><p><br></p><p>되냐???</p><p><br></p><p>ㄱㄱㄱㅋㅋ</p>',
   };
 
   return (
@@ -46,7 +52,7 @@ function DetailPage() {
           height: '100rem',
         }}
       >
-        등록번호 {id} 팀 모집 페이지 내용 ㄱㅁㅇ : 지금은 더미데이터 사용.
+        등록번호 {id} 팀원 모집 페이지 내용 ㄱㅁㅇ : 지금은 더미데이터 사용.
         API요청 받아와서 작업할 예정임.
         <DetailPageBoard detailList={detailList} data={dummydata} />
       </div>
